@@ -1,12 +1,13 @@
-import { reducer as formReducer } from "redux-form";
-import pizzaReducer from "./pizza/reducer";
-import drinkReducer from "./drinks/reducer";
-import createSagaMiddleware from "redux-saga"
-import rootSaga from "./sagas";
-const { createStore, combineReducers, applyMiddleware } = require("redux");
+import { reducer as formReducer } from 'redux-form';
+import createSagaMiddleware from 'redux-saga';
+import pizzaReducer from './pizza/reducer';
+import drinkReducer from './drinks/reducer';
+import rootSaga from './sagas';
 
+// eslint-disable-next-line no-undef
+const { createStore, combineReducers, applyMiddleware } = require('redux');
 
-let reducers = combineReducers({
+const reducers = combineReducers({
   pizzaPage: pizzaReducer,
   drinksPage: drinkReducer,
   form: formReducer,
