@@ -1,10 +1,9 @@
 export const productAPI = {
   getPizzaProduct() {
-    return fetch("/pizzas").then((res) => res.json())
-      
+    return fetch('/pizzas').then((res) => res.json());
   },
 
   getDrinkProduct() {
-    return fetch("/drinks").then((res) => res.json())
+    return fetch('/drinks').then((res) => res.json());
   },
 };
